.artifacts/deploy-prod-snapshot-20260328b/backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
stripe>=8.1.0
websockets>=12.0
supabase>=2.3.0
httpx>=0.26.0
python-multipart>=0.0.7
redis>=5.0.0
python-dotenv>=1.0.1
google-auth>=2.29.0
google-genai>=1.68.0
requests>=2.31.0
h3>=4.4.0
pillow>=10.4.0
slowapi>=0.1.9
prometheus-client>=0.20.0
opentelemetry-api>=1.26.0
opentelemetry-sdk>=1.26.0
opentelemetry-exporter-otlp>=1.26.0
opentelemetry-instrumentation-fastapi>=0.47b0
opentelemetry-instrumentation-httpx>=0.47b0
mem0ai>=0.1.0
